+++ packages/ui/button/src/index.tsx
@@ -9,27 +9,31 @@ const clsPrefix = 'osui-button';
 
 export interface ButtonProps extends Omit<AntdButtonProps, 'type'> {
     type?: ButtonType | 'strong' | 'only-icon';
+    success?: boolean;
+    error?: boolean;
+    warning?: boolean;
 }
 
-/* eslint-disable max-len */
-export interface ButtonInterface extends React.ForwardRefExoticComponent<ButtonProps & React.RefAttributes<HTMLElement>> {
-    Group: typeof AntdButton.Group;
-}
+export type ButtonInterface = React.ForwardRefExoticComponent<ButtonProps & React.RefAttributes<HTMLElement>>;
 
 /* eslint-disable complexity */
-const InternalButton: React.ForwardRefRenderFunction<unknown, ButtonProps> = ({type = 'default', loading, icon, style, disabled, ...props}, ref) => {
+const InternalButton: React.ForwardRefRenderFunction<unknown, ButtonProps> = (
+    {type = 'default', loading, icon, style, disabled, ...props}, ref
+) => {
     let innerIcon = icon;
     let innerStyle = style;
 
-    // 当loading且有icon的button时，icon替换成spinner
+    // 当loading且有icon的button时，icon替换成spinner，不论什么情况都要保持后面的chidlren
     if (loading && icon) {
-        innerIcon = <IconSpinner className={`${clsPrefix}-icon-spinner`} />;
+        innerIcon = <IconSpinner className={`${clsPrefix}-icon-spinner ${clsPrefix}-keep-children`} />;
     }
-    // 当loading但没有icon时，children替换成spinner，且最小宽度保留88
+    // 当loading但没有icon时，children替换成spinner，且最小宽度保留88，根据主题保留或者隐藏children
     if (loading && !icon) {
         innerIcon = <IconSpinner className={`${clsPrefix}-icon-spinner`} />;
         innerStyle = {minWidth: 88, ...style};
     }
+
+    const {success, error, danger, warning} = props;
 
     return (
         <AntdButton
@@ -43,6 +47,9 @@ const InternalButton: React.ForwardRefRenderFunction<unknown, ButtonProps> = ({t
                     {
                         [`${clsPrefix}-loading`]: loading,
                         [`${clsPrefix}-disabled`]: loading || disabled,
+                        [`${clsPrefix}-face-success`]: success,
+                        [`${clsPrefix}-face-error`]: error || danger,
+                        [`${clsPrefix}-face-warning`]: warning,
                     }
                 )
             }
@@ -54,7 +61,5 @@ const InternalButton: React.ForwardRefRenderFunction<unknown, ButtonProps> = ({t
 };
 
 const Button = React.forwardRef<unknown, ButtonProps>(InternalButton) as ButtonInterface;
-
-Button.Group = AntdButton.Group;
 
 export default Button;