+++ packages/ui/button/stories/icloud.stories.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { IconSearch, IconHome } from '@osui/icons';
-import OSUIButton from '../src';
+import Space from '@osui/space';
+import Button from '../src';
 
 export default {
     title: 'Button',
@@ -10,46 +11,92 @@ export const Demo = () => {
     return (
         <>
             <div>
-                <OSUIButton type="primary" style={{ 'margin': '0 20px 20px 0' }}>主要按钮</OSUIButton>
-                <OSUIButton type="primary" loading style={{ 'margin': '0 20px 20px 0' }}>主要按钮</OSUIButton>
-                <OSUIButton type="primary" disabled style={{ 'margin': '0 20px 20px 0' }}>主要按钮</OSUIButton>
+                <Button type="primary" style={{ 'margin': '0 20px 20px 0' }}>主要按钮</Button>
+                <Button type="primary" loading style={{ 'margin': '0 20px 20px 0' }}>主要按钮</Button>
+                <Button type="primary" disabled style={{ 'margin': '0 20px 20px 0' }}>主要按钮</Button>
             </div>
             <div>
-                <OSUIButton type="strong" style={{ 'margin': '0 20px 20px 0' }}>次要按钮</OSUIButton>
-                <OSUIButton type="strong" loading style={{ 'margin': '0 20px 20px 0' }}>次要按钮</OSUIButton>
-                <OSUIButton type="strong" disabled style={{ 'margin': '0 20px 20px 0' }}>次要按钮</OSUIButton>
+                <Button type="strong" style={{ 'margin': '0 20px 20px 0' }}>次要按钮</Button>
+                <Button type="strong" loading style={{ 'margin': '0 20px 20px 0' }}>次要按钮</Button>
+                <Button type="strong" disabled style={{ 'margin': '0 20px 20px 0' }}>次要按钮</Button>
             </div>
             <div>
-                <OSUIButton type="default" style={{ 'margin': '0 20px 20px 0' }}>次要按钮</OSUIButton>
-                <OSUIButton type="default" loading style={{ 'margin': '0 20px 20px 0' }}>次要按钮</OSUIButton>
-                <OSUIButton type="default" disabled style={{ 'margin': '0 20px 20px 0' }}>次要按钮</OSUIButton>
+                <Button type="default" style={{ 'margin': '0 20px 20px 0' }}>次要按钮</Button>
+                <Button type="default" loading style={{ 'margin': '0 20px 20px 0' }}>次要按钮</Button>
+                <Button type="default" disabled style={{ 'margin': '0 20px 20px 0' }}>次要按钮</Button>
             </div>
             <div>
-                <OSUIButton type="link" style={{ 'margin': '0 20px 20px 0' }}>文字按钮</OSUIButton>
-                <OSUIButton type="link" loading style={{ 'margin': '0 20px 20px 0' }}>文字按钮</OSUIButton>
-                <OSUIButton type="link" disabled style={{ 'margin': '0 20px 20px 0' }}>文字按钮</OSUIButton>
+                <Button type="link" style={{ 'margin': '0 20px 20px 0' }}>文字按钮</Button>
+                <Button type="link" loading style={{ 'margin': '0 20px 20px 0' }}>文字按钮</Button>
+                <Button type="link" disabled style={{ 'margin': '0 20px 20px 0' }}>文字按钮</Button>
             </div>
             <div>
-                <OSUIButton icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
-                <OSUIButton disabled icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
-                <OSUIButton loading icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
+                <Button icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
+                <Button loading icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
+                <Button disabled icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
             </div>
             <div>
-                <OSUIButton type="only-icon" icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
-                <OSUIButton type="only-icon" disabled icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
-                <OSUIButton type="only-icon" loading icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
+                <Button type="only-icon" icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
+                <Button type="only-icon" loading icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
+                <Button type="only-icon" disabled icon={<IconSearch />} style={{ 'margin': '0 20px 20px 0' }} />
             </div>
             <div>
-                <OSUIButton type="default" icon={<IconHome />} style={{ 'margin': '0 20px 20px 0' }}>
+                <Button icon={<IconHome />} style={{ 'margin': '0 20px 20px 0' }}>
                     图文按钮
-                </OSUIButton>
-                <OSUIButton type="default" disabled icon={<IconHome />} style={{ 'margin': '0 20px 20px 0' }}>
+                </Button>
+                <Button loading icon={<IconHome />} style={{ 'margin': '0 20px 20px 0' }}>
                     图文按钮
-                </OSUIButton>
-                <OSUIButton type="default" loading icon={<IconHome />} style={{ 'margin': '0 20px 20px 0' }}>
+                </Button>
+                <Button disabled icon={<IconHome />} style={{ 'margin': '0 20px 20px 0' }}>
                     图文按钮
-                </OSUIButton>
+                </Button>
             </div>
+        </>
+    );
+};
+
+export const Group = () => {
+    return (
+        <>
+            <Space>
+                <Button size="small">取消</Button>
+                <Button type="primary" size="small">确定</Button>
+            </Space>
+            <br />
+            <br />
+            <Space>
+                <Button>取消</Button>
+                <Button type="primary">确定</Button>
+            </Space>
+            <br />
+            <br />
+            <Space>
+                <Button size="large">取消</Button>
+                <Button size="large" type="primary">确定</Button>
+            </Space>
+        </>
+    );
+};
+
+export const Face = () => {
+    return (
+        <>
+            <Space>
+                <Button success>成功按钮</Button>
+                <Button success type="primary">成功按钮</Button>
+            </Space>
+            <br />
+            <br />
+            <Space>
+                <Button error>失败按钮</Button>
+                <Button error type="primary">失败按钮</Button>
+            </Space>
+            <br />
+            <br />
+            <Space>
+                <Button warning>失败按钮</Button>
+                <Button warning type="primary">失败按钮</Button>
+            </Space>
         </>
     );
 };