+++ packages/ui/switch/.storybook/preview.js
@@ -6,7 +6,8 @@ import {DocsPage, DocsContainer} from '@storybook/addon-docs/blocks';
 import {
     INITIAL_VIEWPORTS,
   } from '@storybook/addon-viewport';
-import '@osui/theme/dist/theme/vars.css';
+// import '@osui/theme/dist/theme/vars.css';
+import '@osui/icloud-theme/dist/theme/vars.css';
 
 addParameters({
     docs: {